Chapter 클래스의 timestamp 속성의 타입을 int에서 long으로 변경하였습니다.

+++ src/main/java/com/binn328/ym_player/Model/Chapter.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class Chapter {
-    private int timestamp;
+    private long timestamp;
     private String title;
 }
